Fixed main and create mudule validation
